requirements.txt
langchain==0.1.12
langchain-groq==0.0.1
langchain-openai==0.0.8
pinecone-client>=3.0.0
python-dotenv==1.0.1
boto3==1.34.69
python-docx==1.1.0
tiktoken==0.6.0
numpy==1.26.4
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.4
python-multipart==0.0.9
jinja2==3.1.3
